Add CoreLib shared source project (#244)
* Add CoreLib shared source project

Add a CoreLib shared source project (shproj) to allow displaying shared
source files. With the move of the shared folder to a different location
VS isn't able to display the sources anymore. Adding a link attribute to
the compile items in the projitems fixes this but the information is
lost after a project reload. Therefore adding a shared project to the
solution.

* Update System.Private.CoreLib.Shared.shproj

* Update System.Private.CoreLib.Shared.shproj