Fix all GC stress failures in Pri1 tests on Unix with crossgen2 (#1619)
* Fix all GC stress failures in Pri1 tests on Unix with crossgen2

Without this PR, 90% of the CoreCLR Pri1 tests are failing due to GC
holes. All of the errors are caused by various bugs in the GCRefMap
generation that this PR fixes. With this PR, all the tests that pass
without GC stress also pass with it.

1. GetCorElementType() was incorrectly casting from TypeFlags to
CoreElementType. It was checking for a range of TypeFlags, but that
range didn't map to CorElementType 1:1. That resulted e.g. in getting
ELEMENT_TYPE_R4 for TypeFlags.IntPtr.
2. Besides that, the GetCorElementType() was also incorrectly mapping
Enum to ELEMENT_TYPE_CLASS.
3. The ReportPointersFromStructInRegisters function was incorrectly
assuming that the false stack entries correspond to stack slots while
they in fact correspond to bytes.
4. Arguments that are passed on stack were incorrectly reported as being
passed in registers (if there were still any registers available)
5. Members of structs embedded in other structs were reporting wrong
offsets (we were missing updating the offset by the delta parameter in
the GcScanValueType when calling GCScanRoots)
6. The code that determines if return values are passed in registers or
on stack was completely missing handling of structs returned in
registers.

* Fix typo in comment