Include apphost_version parameter in download URLs for both GUI and CLI apps
So far the code only included the `&apphost_version=<version>` URL parameter in the URL used by the error dialog in GUI apps. The error written to stderr/trace would not have that parameter in its URL.

This change include this parameter in all cases. This also means that comhost/ijwhost and so on will also include the `apphost_version` parameter in their error codes and it will be the version of the comhost/ijwhost and so on. This is not 100% correct, but for the purposed of the website this should be enough (for now we don't see a reason to differentiate between the various hosting scenarios).

Adds a new test to cover this case.