Warn if DynamicallyAccessedMembersAttribute is used directly on a method (dotnet/linker#1247)
We had to allow the attribute on a method so that we could annotate System.Type instance methods. In that case the attribute on the method means "annotate the this parameter with this".

But using the attribute anywhere else is not allowed. Linker should warn about this.

Commit migrated from https://github.com/dotnet/linker/commit/455ceb827d0ce95dce523ed6ba3e19dcb4d14de5