Unix: add mode to inline Socket continuations (#37974)
* Unix: add mode to inline Socket continuations

On Unix, socket continuations are dispatched to the ThreadPool
from an event thread. This avoids continuations blocking the
event handling.

This adds an option to disable that dispatch. Continuations for
socket operations will be executed on the event thread directly.
This removes the overhead of context switching to the ThreadPool.

Currently this is implemented as an application level switch for
benchmarking, and experimentation. It may be made controllable
at the Socket level.

To avoid the event threads being a bottleneck, ProcessorCount
event threads are created.

* Add assert to verify Error is filtered out

* Add internal Socket.PreferInlineCompletions

* Add envvar to control PreferInlineCompletions default

* PR feedback

* Squash DOTNET_SYSTEM_NET_SOCKETS_INLINE_COMPLETIONS envvars

* Don't use Unix specific SocketAsyncEngine in common SafeHandle code

* PR feedback