Consider that retbuf arg can point to GC heap in fgCreateCallDispatcherAndGetResult() (#39815)
Caller return buffer argument can point to GC heap while DispatchTailCalls expects the return value argument to point to the stack. We should use a temporary stack allocated return buffer to hold the value during the dispatcher call and copy the value back to the caller return buffer after that.