Limit Debugger.Log() calls in EventSource (#40823)
* Limit Debugger.Log() calls in EventSource

Fixing #40799 

Previously we would send a message via Debugger.Log() for event call to ReportOutOfBandMessage().
Now we only do it the first 16 times, then all errors after that are silent. This should alleviate performance
overhead from a scenarios with a large number of errors.

I made this change quickly via github to help out Bing. I have never compiled or tested it prior to the PR
but fingers crossed it appears simple enough that it has a decent chance to work : )

* Remove trailing whitespace

Co-authored-by: David Mason <davmason@microsoft.com>